* commit '67f72d8a0f3bbdf843a913a9f724811c42eb3a7d':
  Only force scroll to bottom in active convo view when the convo was just opened freshly. When actually reading messages above, don't scroll down when receiving new ones.
  Fixed weird behaviour. Don't even know how to describe this... ffs
  Update convo name label on changed metadata
  Added safety check for when writing messages out to disk
  Moved request authentication form to view top (edit convo metadata dialog)
  When pulling messages, convo metadata is pulled first and updated if something changed.

+++ GlitchedEpistle.Client.Windows/ViewModels/UserControls/ActiveConvoViewModel.cs
@@ -75,6 +75,13 @@ namespace GlitchedPolygons.GlitchedEpistle.Client.Windows.ViewModels.UserControl
             set => Set(ref pulling, value);
         }
 
+        private string name;
+        public string Name
+        {
+            get => name;
+            set => Set(ref name, value);
+        }
+
         private Visibility clipboardTickVisibility = Visibility.Hidden;
         public Visibility ClipboardTickVisibility
         {
@@ -114,6 +121,7 @@ namespace GlitchedPolygons.GlitchedEpistle.Client.Windows.ViewModels.UserControl
                 StopAutomaticPulling();
                 Messages = new ObservableCollection<MessageViewModel>();
                 activeConvo = value;
+                Name = value?.Name;
                 Task.Run(() =>
                 {
                     LoadLocalMessages();
@@ -146,7 +154,10 @@ namespace GlitchedPolygons.GlitchedEpistle.Client.Windows.ViewModels.UserControl
             SendFileCommand = new DelegateCommand(OnSendFile);
             PressedEscapeCommand = new DelegateCommand(OnPressedEscape);
             CopyConvoIdToClipboardCommand = new DelegateCommand(OnClickedCopyConvoIdToClipboard);
+
             eventAggregator.GetEvent<LogoutEvent>().Subscribe(StopAutomaticPulling);
+            eventAggregator.GetEvent<ChangedConvoMetadataEvent>().Subscribe(OnChangedConvoMetadata);
+
             settings.Load();
         }
 
@@ -170,6 +181,16 @@ namespace GlitchedPolygons.GlitchedEpistle.Client.Windows.ViewModels.UserControl
         {
             StopAutomaticPulling();
             scheduledUpdateRoutine = methodQ.Schedule(PullNewestMessages, MSG_PULL_FREQUENCY);
+        }
+
+        private void OnChangedConvoMetadata(string convoId)
+        {
+            var convo = convoProvider[convoId];
+            if (convo is null)
+            {
+                return;
+            }
+            Name = convo.Name;
         }
 
         private void TransferQueuedMessagesToUI()
@@ -335,10 +356,14 @@ namespace GlitchedPolygons.GlitchedEpistle.Client.Windows.ViewModels.UserControl
 
         private void WriteMessageToDisk(Message message)
         {
-            File.WriteAllText(
-                contents: JsonConvert.SerializeObject(message),
-                path: Path.Combine(Paths.CONVOS_DIRECTORY, ActiveConvo.Id, message.TimestampUTC.ToString("yyyyMMddHHmmssfff"))
-            );
+            string path = Path.Combine(Paths.CONVOS_DIRECTORY, ActiveConvo.Id, message.TimestampUTC.ToString("yyyyMMddHHmmssfff"));
+            if (!File.Exists(path))
+            {
+                File.WriteAllText(
+                    contents: JsonConvert.SerializeObject(message),
+                    path: path
+                );
+            }
         }
 
         private void PullNewestMessages()
@@ -352,6 +377,33 @@ namespace GlitchedPolygons.GlitchedEpistle.Client.Windows.ViewModels.UserControl
 
             Task.Run(async () =>
             {
+                // Pull convo metadata first.
+                var convo = convoProvider[ActiveConvo.Id];
+                var metadata = await convoService.GetConvoMetadata(ActiveConvo.Id, ActiveConvo.PasswordSHA512, user.Id, user.Token.Item2);
+
+                if (metadata != null && !convo.Equals(metadata))
+                {
+                    convo.Name = ActiveConvo.Name = metadata.Name;
+                    convo.CreatorId = ActiveConvo.CreatorId = metadata.CreatorId;
+                    convo.Description = ActiveConvo.Description = metadata.Description;
+                    convo.ExpirationUTC = ActiveConvo.ExpirationUTC = metadata.ExpirationUTC;
+                    convo.CreationTimestampUTC = ActiveConvo.CreationTimestampUTC = metadata.CreationTimestampUTC;
+                    convo.BannedUsers = ActiveConvo.BannedUsers = metadata.BannedUsers.Split(',').ToList();
+                    convo.Participants = ActiveConvo.Participants = metadata.Participants.Split(',').ToList();
+
+                    var _convo = convoProvider[convo.Id];
+                    if (_convo != null)
+                    {
+                        convoProvider.Convos.Remove(_convo);
+                    }
+
+                    convoProvider.Convos.Add(convo);
+                    convoProvider.Save();
+
+                    Application.Current?.Dispatcher?.Invoke(() => eventAggregator.GetEvent<ChangedConvoMetadataEvent>().Publish(convo.Id));
+                }
+
+                // Pull newest messages.
                 int i = 0;
                 if (Messages.Count > 0)
                 {
